requirements.txt
# FlexiFinance - Django MicroFinance Platform
# Core Django Framework
Django==5.2.8

# Database
psycopg2-binary==2.9.9  # PostgreSQL adapter
PyMySQL==1.1.0          # Pure Python MySQL adapter (no system deps)
dj-database-url==2.1.0  # Database URL parsing

# API and Serialization
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1

# Authentication and Security
django-allauth==65.0.2
django-extensions==3.2.3
django-crispy-forms==2.3  # Form rendering
crispy-bootstrap5==2025.6  # Latest version compatible with Python 3.12

# File Handling and Storage
django-storages==1.14.4  # Cloud storage backends
Pillow==11.0.0          # Image processing (Python 3.12 compatible)
boto3==1.34.63          # AWS S3 integration

# Payment Processing
requests==2.31.0        # HTTP requests for M-Pesa API
mpesa-api==0.2.3        # M-Pesa SDK (latest available version)
stripe==6.7.0           # Stripe payment processing

# Task Queue and Caching
celery==5.3.6           # Asynchronous task processing
redis==5.0.1            # Redis client for caching
django-redis==5.4.0     # Django Redis integration

# Email and Notifications
django-ses==3.5.0       # AWS SES for email
sendgrid-django==4.2.0  # SendGrid integration (latest available version)

# Environment and Configuration
python-decouple==3.8    # Environment variable management
python-dotenv==1.0.0    # .env file support
django-environ==0.11.2  # Environment configuration

# HTTP and API Clients
httpx==0.26.0           # HTTP client
urllib3==2.2.1          # URL handling

# Data Validation and Parsing
Pydantic==2.6.0         # Data validation
marshmallow==3.20.2     # Object serialization

# Business Logic and Calculations
numpy==1.26.3           # Numerical computations
pandas==2.2.0           # Data analysis

# Security
cryptography==42.0.2    # Cryptographic functions
django-csp==3.7         # Content Security Policy (Python 3.12 compatible)

# Monitoring and Logging
django-silk==5.1.0      # Profiling and debugging (latest for Python 3.12)
django-debug-toolbar     # Debug toolbar (auto-resolve compatible version)

# Testing
pytest==8.0.0           # Testing framework
pytest-django==4.8.0    # Django testing integration
factory-boy==3.3.0      # Test data generation
coverage==7.4.3         # Code coverage
faker==22.2.0           # Fake data generation

# Development Tools
flake8==7.0.0           # Code linting
black==24.2.0           # Code formatting
isort==5.13.2           # Import sorting
pre-commit==3.6.0       # Pre-commit hooks

# Documentation
Sphinx==7.2.6           # Documentation generation
sphinx-rtd-theme==1.3.0 # Documentation theme

# Static Files and Media
whitenoise==6.6.0       # Static file serving
django-compressor==4.4  # Static file compression (Python 3.12 compatible)

# Forms and Validation
django-formtools==2.5.1 # Form utilities (latest in 2.5 series)
django-choices==1.7.2   # Choice field utilities
django-widget-tweaks==1.5.0  # Form field customization


# Pagination and Filtering
django-filter==24.2     # Filtering functionality (Python 3.12 compatible)
django-pagination==1.0.7 # Pagination utilities (latest available version)

# Internationalization
django-rosetta==0.9.8   # Translation management
django-multiselectfield==0.1.13  # Multi-select fields

# Data Import/Export
django-import-export==4.1.1  # Data import/export
tablib==3.5.0           # Tabular data library (compatible with django-import-export)

# Background Tasks
django-celery-beat      # Celery beat scheduler (auto-resolve compatible version)
django-celery-results   # Celery results backend (auto-resolve compatible version)

# API Documentation
drf-spectacular==0.27.1  # API documentation
swagger-spec-validator==3.0.3  # Swagger validation

# Development Dependencies
ipython==8.28.0         # Enhanced Python shell (latest for Python 3.12)
jupyter==1.0.0          # Jupyter notebook support
django-debug-toolbar-line-profiler==0.6.1  # Line profiling (latest available version)

# Production Dependencies
gunicorn==21.2.0        # WSGI HTTP Server
django-cachalot          # Database caching (auto-resolve compatible version)

# Additional Utilities
python-dateutil==2.9.0  # Date utilities
pytz==2024.1            # Timezone handling
beautifulsoup4==4.12.3  # HTML/XML parsing
lxml==5.3.0             # XML processing (Python 3.12 compatible)

# Financial Calculations
numpy-financial==1.0.0   # Financial calculations (numpy-financial package)
yfinance==0.2.28        # Financial data (if needed)

# Communication Services
twilio==9.0.5           # SMS and voice services
django-notifications-hq==1.8.3  # Notification system (latest available version)

# Analytics and Tracking
django-analytical==3.2.0  # Web analytics integration (latest available version)
